+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 
 ### Breaking changes
 - The `tangent1` and `tangent2` properties of `HermiteKeyframe`s have been renamed to `t1` and `t2` respectively in order to shorten the JSON form of these keyframes. This should allow these keyframes to be inlined more often, which should make it easier to read the JSON for humans.
+- The `unk_sdt_f2_32` property in the following actions has been changed to a boolean and the documentation has been updated:
+  - `PointSprite`
+  - `Line`
+  - `QuadLine`
+  - `BillboardEx`
+  - `MultiTextureBillboardEx`
+  - `Model`
+  - `LegacyTracer`
+  - `GPUStandardParticle`
+  - `GPUStandardCorrectParticle`
+  - `Tracer`
 
 ### Improvements and new features
 - The parsing functions have been updated to also support reading FXRs that have been extracted from the game's memory.